ga_mcp_py/pyproject.toml
[project]
name = "ga-mcp-py"
version = "0.1.0"
description = "Google Analytics MCP (Model Context Protocol) Python implementation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client",
    "oauth2client",
    "mcp[cli]",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "main.py",
    "README.md",
    "pyproject.toml"
]
